5th_Draft/requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
curl_cffi==0.13.0
Flask==3.1.2
frozendict==2.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
multitasking==0.0.12
nltk==3.9.1
numpy==2.3.2
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
protobuf==5.29.5
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.7.34
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.65
python-dotenv==1.0.1
gunicorn==22.0.0
matplotlib==3.9.2
rapidfuzz==3.10.1
openai==1.51.2
google-generativeai==0.8.3
